ADD: basic support for RFC4648 version of base64
As of this commit, it is necessary to edit src/main.c and recompile the program to
switch from one RFC to the other. The ability to select it at runtime via a flag on
the command line will be added later.

Note: (Reminder) RFC4648 simply replaces the characters '+' and '/' with '-' and '_'
respectively in order to keep the base64 output URL friendly.